move filtering token to under property filter
