+++ src/property-filter/filtering-token/styles.scss
@@ -3,10 +3,11 @@
  SPDX-License-Identifier: Apache-2.0
 */
 
-@use '../../styles' as styles;
-@use '../../styles/tokens' as awsui;
 @use '@cloudscape-design/component-toolkit/internal/focus-visible' as focus-visible;
-@use '../../../token-group/constants' as constants;
+@use '../../internal/styles' as styles;
+@use '../../internal/styles/tokens' as awsui;
+@use '../../internal/hooks/focus-visible' as focus-visible;
+@use '../../token-group/constants' as constants;
 
 .root {
   display: flex;
